BugID:16959715: reduce codesize for LoRaWAN 4.4.2

+++ board/eml3047/src/hw_rtc.c
@@ -707,7 +707,7 @@ TimerTime_t HW_RTC_TempCompensation( TimerTime_t period, float temperature )
     interim = ( ( float ) period * ppm ) / 1e6;
     // Calculate the resulting time period
     interim += period;
-    interim = floor( interim );
+    interim = floorf( interim );
 
     if( interim < 0.0 )
     {
